+++ stream-chat-android-state/src/test/java/io/getstream/chat/android/offline/querychannels/DefaultChatEventHandlerTest.kt
@@ -46,7 +46,7 @@ internal class DefaultChatEventHandlerTest {
         val clientState = mock<ClientState> {
             whenever(it.user) doReturn MutableStateFlow(currentUser)
         }
-        val eventHandler = DefaultChatEventHandler(MutableStateFlow(listOf(channel)), clientState)
+        val eventHandler = DefaultChatEventHandler(MutableStateFlow(mapOf(channel.cid to channel)), clientState)
         val event = randomMemberRemovedEvent(channel.cid, member)
 
         val result = eventHandler.handleChatEvent(event = event, filter = Filters.neutral(), cachedChannel = null)
@@ -60,7 +60,7 @@ internal class DefaultChatEventHandlerTest {
         val clientState = mock<ClientState> {
             whenever(it.user) doReturn MutableStateFlow(randomUser())
         }
-        val eventHandler = DefaultChatEventHandler(MutableStateFlow(listOf(channel)), clientState)
+        val eventHandler = DefaultChatEventHandler(MutableStateFlow(mapOf(channel.cid to channel)), clientState)
         val event = randomMemberRemovedEvent(channel.cid)
 
         val result = eventHandler.handleChatEvent(event = event, filter = Filters.neutral(), cachedChannel = null)
@@ -73,7 +73,7 @@ internal class DefaultChatEventHandlerTest {
         val clientState = mock<ClientState> {
             whenever(it.user) doReturn MutableStateFlow(null)
         }
-        val eventHandler = DefaultChatEventHandler(MutableStateFlow(emptyList()), clientState)
+        val eventHandler = DefaultChatEventHandler(MutableStateFlow(emptyMap()), clientState)
         val event = randomMemberRemovedEvent()
 
         val result = eventHandler.handleChatEvent(event = event, filter = Filters.neutral(), cachedChannel = null)
@@ -86,7 +86,7 @@ internal class DefaultChatEventHandlerTest {
         val clientState = mock<ClientState> {
             whenever(it.user) doReturn MutableStateFlow(null)
         }
-        val eventHandler = DefaultChatEventHandler(MutableStateFlow(emptyList()), clientState)
+        val eventHandler = DefaultChatEventHandler(MutableStateFlow(emptyMap()), clientState)
         val event = randomMemberAddedEvent()
 
         val result = eventHandler.handleChatEvent(event = event, filter = Filters.neutral(), cachedChannel = null)
@@ -99,7 +99,7 @@ internal class DefaultChatEventHandlerTest {
         val clientState = mock<ClientState> {
             whenever(it.user) doReturn MutableStateFlow(randomUser())
         }
-        val eventHandler = DefaultChatEventHandler(MutableStateFlow(emptyList()), clientState)
+        val eventHandler = DefaultChatEventHandler(MutableStateFlow(emptyMap()), clientState)
         val event = randomMemberAddedEvent()
 
         val result = eventHandler.handleChatEvent(event = event, filter = Filters.neutral(), cachedChannel = null)
@@ -114,7 +114,7 @@ internal class DefaultChatEventHandlerTest {
         val clientState = mock<ClientState> {
             whenever(it.user) doReturn MutableStateFlow(currentUser)
         }
-        val eventHandler = DefaultChatEventHandler(MutableStateFlow(emptyList()), clientState)
+        val eventHandler = DefaultChatEventHandler(MutableStateFlow(emptyMap()), clientState)
         val event = randomMemberAddedEvent(cid = channel.cid, member = randomMember(user = currentUser))
 
         val result = eventHandler.handleChatEvent(event = event, filter = Filters.neutral(), cachedChannel = channel)
@@ -126,7 +126,7 @@ internal class DefaultChatEventHandlerTest {
     fun `When received NotificationMessageNewEvent Should watch and add the channel`() {
         val cid = randomCID()
         val clientState = mock<ClientState>()
-        val eventHandler = DefaultChatEventHandler(MutableStateFlow(emptyList()), clientState)
+        val eventHandler = DefaultChatEventHandler(MutableStateFlow(emptyMap()), clientState)
         val event = randomNotificationMessageNewEvent(cid = cid)
 
         val result = eventHandler.handleChatEvent(event = event, filter = Filters.neutral(), cachedChannel = null)
@@ -138,7 +138,7 @@ internal class DefaultChatEventHandlerTest {
     fun `When received NotificationAddedToChannelEvent Should watch and add the channel`() {
         val cid = randomCID()
         val clientState = mock<ClientState>()
-        val eventHandler = DefaultChatEventHandler(MutableStateFlow(emptyList()), clientState)
+        val eventHandler = DefaultChatEventHandler(MutableStateFlow(emptyMap()), clientState)
         val event = randomNotificationAddedToChannelEvent(cid = cid)
 
         val result = eventHandler.handleChatEvent(event = event, filter = Filters.neutral(), cachedChannel = null)
@@ -154,7 +154,7 @@ internal class DefaultChatEventHandlerTest {
         val clientState = mock<ClientState> {
             whenever(it.user) doReturn MutableStateFlow(currentUser)
         }
-        val eventHandler = DefaultChatEventHandler(MutableStateFlow(listOf(channel)), clientState)
+        val eventHandler = DefaultChatEventHandler(MutableStateFlow(mapOf(channel.cid to channel)), clientState)
         val event = randomNotificationRemovedFromChannelEvent(channel.cid, member = member)
 
         val result = eventHandler.handleChatEvent(event = event, filter = Filters.neutral(), cachedChannel = null)
@@ -168,7 +168,7 @@ internal class DefaultChatEventHandlerTest {
         val clientState = mock<ClientState> {
             whenever(it.user) doReturn MutableStateFlow(randomUser())
         }
-        val eventHandler = DefaultChatEventHandler(MutableStateFlow(listOf(channel)), clientState)
+        val eventHandler = DefaultChatEventHandler(MutableStateFlow(mapOf(channel.cid to channel)), clientState)
         val event = randomNotificationRemovedFromChannelEvent(channel.cid)
 
         val result = eventHandler.handleChatEvent(event = event, filter = Filters.neutral(), cachedChannel = null)
@@ -181,7 +181,7 @@ internal class DefaultChatEventHandlerTest {
         val clientState = mock<ClientState> {
             whenever(it.user) doReturn MutableStateFlow(null)
         }
-        val eventHandler = DefaultChatEventHandler(MutableStateFlow(emptyList()), clientState)
+        val eventHandler = DefaultChatEventHandler(MutableStateFlow(emptyMap()), clientState)
         val event = randomNotificationRemovedFromChannelEvent()
 
         val result = eventHandler.handleChatEvent(event = event, filter = Filters.neutral(), cachedChannel = null)